requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
openai==1.14.0
anthropic==0.23.1
google-generativeai==0.3.0
sentence-transformers==2.7.0
scikit-learn==1.4.1.post1
wikipedia-api==0.6.0
transformers==4.40.1
nltk==3.8.1
spacy==3.7.4
en_core_web_sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl